Initial binreloc modernization
Original binreloc interfaces to return "bindir" etc assume a pure
posix style installation hierarchy. This may not be the case, and
want to leave interpretation of paths from exe to other locations up
to developer.

Copy original binreloc files to "binreloc.EXT" for reference. Strip
down hsf_binreloc copies to minimum interface needed to initialize and
return the full path to the exe.